+++ src/pages/SignUpPage/hooks/useSubmit.ts
@@ -1,9 +1,16 @@
-import { Dispatch, FormEvent, SetStateAction } from 'react';
+import {
+  Dispatch,
+  FormEvent,
+  SetStateAction,
+  useEffect,
+  useState,
+} from 'react';
 import { useFetchSignUp } from '@apis/auth';
 
 interface useSubmitProps {
   email: string;
   password: string;
+  passwordConfirm: string;
   fullName: string;
   isAllPassed: boolean;
   setIsSubmitted: Dispatch<SetStateAction<boolean>>;
@@ -12,11 +19,22 @@ interface useSubmitProps {
 const useSubmit = ({
   email,
   password,
+  passwordConfirm,
   fullName,
   isAllPassed,
   setIsSubmitted,
 }: useSubmitProps) => {
+  const [isSignUpDisabled, setIsSignUpDisabled] = useState(true);
   const { signUpMutate, isSignUpSuccess, isSignUpError } = useFetchSignUp();
+
+  useEffect(() => {
+    email.length > 5 &&
+    password.length > 4 &&
+    passwordConfirm.length > 4 &&
+    fullName.length > 2
+      ? setIsSignUpDisabled(false)
+      : setIsSignUpDisabled(true);
+  }, [email, password, passwordConfirm, fullName]);
 
   const handleSubmit = (e: FormEvent) => {
     e.preventDefault();
@@ -30,7 +48,7 @@ const useSubmit = ({
       });
   };
 
-  return { handleSubmit, isSignUpError, isSignUpSuccess };
+  return { isSignUpDisabled, handleSubmit, isSignUpError, isSignUpSuccess };
 };
 
 export default useSubmit;
